Added the Pause and Next Frame functions
In an attempt to make them not crash (as often) as the same function in SF4 Camera Man, this will only work if the user has not paused the game himself. Once the user has paused the game it will not function anymore and the user will have to re-enter the stage.